VP-1854: Move common components to root "components" folder
