+++ client-app/src/components/shopping-cart/cart-button/index.vue
@@ -1,9 +1,9 @@
 <template>
-  <button class="btn btn-secondary p-2 m-10"
+  <button class="btn btn-secondary p-2"
           @click="onClick">
     <i class="fas fa-shopping-cart"></i> <span class="badge badge-info">{{ itemsQuantity >0? itemsQuantity : "" }}</span>
   </button>
 </template>
 
-<script lang="ts" src="./cart-component.ts"></script>
+<script lang="ts" src="./cart-button.ts"></script>
