+++ client-app/src/components/shopping-cart/cart-item/index.vue
