+++ client-app/src/components/shopping-cart/cart-items-list/cart-items-list.ts
@@ -1,7 +1,7 @@
 import Vue from "vue";
 import Component from "vue-class-component";
 import { Prop, Emit } from "vue-property-decorator";
-import CartItem from "@catalog/components/cart-item/index.vue";
+import CartItem from "components/shopping-cart/cart-item/index.vue";
 import { CartLineItem } from '@common/api/api-clients';
