+++ client-app/src/components/shopping-cart/cart-items-list/index.vue
