+++ client-app/src/components/shopping-cart/cart-summary/index.vue
