+++ client-app/src/pages/catalog/main.ts
@@ -1,8 +1,8 @@
 import Vue from "vue";
 import i18n from "@i18n";
+import AddToCartButton from "components/shopping-cart/add-to-cart-button/index.vue";
+import CartButton from "components/shopping-cart/cart-button/index.vue";
 import store from "store";
-import AddToCartButton from "@catalog/components/add-to-cart-button/index.vue";
-import CartComponent from "@catalog/components/cart-component/index.vue";
 import InitializationService from "@common/services/initialization.service";
 import ActiveOrderSidebar from './components/active-order-sidebar/index.vue';
 import cartModule from "./store/modules/cart";
@@ -31,7 +31,7 @@ for(const addToCartEl of addToCartElements){
 new Vue({
   i18n,
   store,
-  render: h => h(CartComponent)
+  render: h => h(CartButton)
 }).$mount("#app-cart", true);
